+++ packages/zenrouter_file_annotation/lib/zenrouter_file_annotation.dart
@@ -5,3 +5,5 @@ export 'src/analyzer/path_parser.dart';
 export 'src/analyzer/route_element.dart';
 export 'src/analyzer/layout_element.dart';
 export 'src/analyzer/route_tree.dart';
+export 'src/codegen/layout_code_generator.dart';
+export 'src/codegen/route_code_generator.dart';